更新部分icon；添加loanProduct最新label；change TabLoan moneyfilter from double to int;

+++ app/src/main/java/com/beihui/market/ui/contract/TabLoanContract.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface TabLoanContract {
 
     interface Presenter extends BasePresenter {
-        void filterAmount(double amount);
+        void filterAmount(int amount);
 
         void filterDueTime(int selected);
